Move storage header impl into cpp
