+++ bcos-table/src/KeyPageStorage.cpp
@@ -26,6 +26,34 @@
 
 namespace bcos::storage
 {
+KeyPageStorage::KeyPageStorage(std::shared_ptr<StorageInterface> _prev, bool setRowWithDirtyFlag,
+    size_t _pageSize, uint32_t _blockVersion,
+    std::shared_ptr<const std::set<std::string, std::less<>>> _ignoreTables,
+    bool _ignoreNotExist)
+  : storage::StateStorageInterface(std::move(_prev)),
+    m_blockVersion(_blockVersion),
+    m_pageSize(_pageSize > MIN_PAGE_SIZE ? _pageSize : MIN_PAGE_SIZE),
+    m_splitSize(m_pageSize / 3 * 2),
+    m_mergeSize(m_pageSize / 4),
+    m_buckets(std::max(1u, std::thread::hardware_concurrency())),
+    m_ignoreTables(std::move(_ignoreTables)),
+    m_ignoreNotExist(_ignoreNotExist),
+    m_setRowWithDirtyFlag(setRowWithDirtyFlag)
+{
+    if (!m_ignoreTables)
+    {
+        auto ignore = std::make_shared<std::set<std::string, std::less<>>>();
+        ignore->insert(std::string(SYS_TABLES));
+        m_ignoreTables = ignore;
+    }
+}
+
+KeyPageStorage::~KeyPageStorage()
+{
+    m_recoder.clear();
+    m_buckets.clear();
+}
+
 void KeyPageStorage::asyncGetPrimaryKeys(std::string_view tableView,
     const std::optional<storage::Condition const>& _condition,
     std::function<void(Error::UniquePtr, std::vector<std::string>)> _callback)
